Part 04 - Exercise 4.21

+++ controllers/blogs.js
@@ -31,10 +31,17 @@ blogsRouter.post('/', tokenExtractor, async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', tokenExtractor, async (request, response) => {
 
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user.toString() === request.userId.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).end()
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
